* CacheableMap Serializer and Deserializer are now inner classes
