+++ src/integration/java/com/couchbase/client/spring/cache/wiring/javaConfig/CacheEnabledTestConfiguration.java
@@ -18,11 +18,8 @@ package com.couchbase.client.spring.cache.wiring.javaConfig;
 import com.couchbase.client.spring.cache.CouchbaseCacheManager;
 import com.couchbase.client.spring.cache.CacheBuilder;
 import com.couchbase.client.spring.cache.TestConfiguration;
-import com.couchbase.client.spring.cache.wiring.CacheableMap;
-import com.couchbase.client.spring.cache.wiring.CacheableMapSerializer;
 import com.couchbase.client.spring.cache.wiring.CachedService;
 
-import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.rh.rhapsody.commons.deser.jackson.SafeObjectMapper;
 import org.springframework.cache.CacheManager;
 import org.springframework.cache.annotation.EnableCaching;
